+++ CHANGES.md
@@ -8,6 +8,7 @@
  * [METAMODEL-1140] - Allowed SalesforceDataContext without a security token.
  * [METAMODEL-1141] - Added RFC 4180 compliant CSV parsing.
  * [METAMODEL-1144] - Optimized evaluation of conditional client-side JOIN statements.
+ * [METAMODEL-1145] - Fixed bug with modelling JDBC table relationships when there are multiple keys involved in the relationship.
 
 ### Apache MetaModel 4.6.0